+++ NEWS.md
@@ -1,3 +1,13 @@
+# clinpubr 1.2.0
+
+* Added `data_overview()` to get a quick overview of a data frame. It includes diagnostic information about common data issues and gives recommendations for data cleaning.
+
+* Added several outlier detection methods: `mad_outlier()`, `zscore_outlier()`, `iqr_outlier()`.
+
+* Added `merge_by_substring()` to merge data frames by substring matching.
+
+* Added `recursive` argument to `combine_files()`.
+
 # clinpubr 1.1.1
 
 * Fixed PR plot visual inconsistency in `classif_model_compare()`.